+++ Package.swift
@@ -22,6 +22,7 @@ let package = Package(
         // .package(url: /* package url */, from: "1.0.0"),
 //        .package(path: "../../Projects/Tarot"),
         .package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
+        .package(url: "https://github.com/apple/swift-system", from: "1.0.0"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -34,12 +35,14 @@ let package = Package(
         .target(
             name: "Graph",
             dependencies: [
+                .product(name: "SystemPackage", package: "swift-system"),
             ]),
         .executableTarget(
             name: "FlowTool",
             dependencies: [
                 "Flows",
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
+                .product(name: "SystemPackage", package: "swift-system"),
             ]),
         .testTarget(
             name: "FlowsTests",