+++ Sources/Graph/Constraints/Constraints.swift
@@ -5,8 +5,6 @@
 //  Created by Stefan Urbanek on 13/06/2022.
 //
 
-import Foundation
-
 /*
 
  Potential merge: