+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+## New Version
+
+(probably 0.1.1)
+
+- removed dependency on Foundation and replaced with [Swift System](https://github.com/apple/swift-system)
+
 
 ## Version 0.1